+++ sdk/src/test/java/co/omise/android/TokenizationMethodTest.kt
@@ -2,11 +2,8 @@ package co.omise.android
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import co.omise.android.models.Capability
-import co.omise.android.models.Googlepay
 import co.omise.android.models.PaymentMethod
 import co.omise.android.models.TokenizationMethod
-import com.fasterxml.jackson.databind.ObjectMapper
-import org.json.JSONObject
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
@@ -55,127 +52,5 @@ class TokenizationMethodTest {
         val tokenizationMethod = TokenizationMethod.creator("aaa")
 
         assertEquals(TokenizationMethod.Unknown("aaa"), tokenizationMethod)
-    }
-
-    //
-    // Google Pay Tests
-    //
-
-    @Test
-    fun googlepay_isReadyToPayRequest_returnsCorrectValues() {
-        val googlepay = Googlepay("pkey_123", arrayListOf("Visa"), 3000, "thb", "merchantId")
-        val isReadyToPay = googlepay.isReadyToPayRequest()
-
-
-        val expected = JSONObject(
-            """
-            {
-              "apiVersion": 2,
-              "apiVersionMinor": 0,
-              "allowedPaymentMethods": [
-                {
-                  "type": "CARD",
-                  "parameters": {
-                    "allowedAuthMethods": [
-                      "PAN_ONLY"
-                    ],
-                    "allowedCardNetworks": [
-                      "VISA"
-                    ]
-                  }
-                }
-              ]
-            }
-            """
-        )
-
-        val mapper = ObjectMapper()
-        assertEquals(mapper.readTree(expected.toString()), mapper.readTree(isReadyToPay.toString()))
-    }
-
-
-    @Test
-    fun googlepay_useDefaultCardNetworksIfMissingFromCapabilities() {
-        val googlepay = Googlepay("pkey_123", arrayListOf(), 3000, "thb", "merchantId")
-        val isReadyToPay = googlepay.isReadyToPayRequest()
-
-
-        val expected = JSONObject(
-            """
-            {
-              "apiVersion": 2,
-              "apiVersionMinor": 0,
-              "allowedPaymentMethods": [
-                {
-                  "type": "CARD",
-                  "parameters": {
-                    "allowedAuthMethods": [
-                      "PAN_ONLY"
-                    ],
-                    "allowedCardNetworks": [
-                      "AMEX",
-                      "JCB",
-                      "MASTERCARD",
-                      "VISA"
-                    ]
-                  }
-                }
-              ]
-            }
-            """
-        )
-
-        val mapper = ObjectMapper()
-        assertEquals(mapper.readTree(expected.toString()), mapper.readTree(isReadyToPay.toString()))
-    }
-
-
-    @Test
-    fun googlepay_getPaymentDataRequest_returnsCorrectValues() {
-        val googlepay = Googlepay("pkey_123", arrayListOf("American Express", "MasterCard", "JCB"), 3000, "sgd", "merchantId")
-        val getPaymentDataRequest = googlepay.getPaymentDataRequest()
-
-
-        val expected = JSONObject(
-            """
-            {
-              "apiVersion": 2,
-              "apiVersionMinor": 0,
-              "allowedPaymentMethods": [
-                {
-                  "type": "CARD",
-                  "parameters": {
-                    "allowedAuthMethods": [
-                      "PAN_ONLY"
-                    ],
-                    "allowedCardNetworks": [
-                      "AMEX",
-                      "MASTERCARD",
-                      "JCB"
-                    ]
-                  },
-                  "tokenizationSpecification": {
-                    "type": "PAYMENT_GATEWAY",
-                    "parameters": {
-                      "gateway": "omise",
-                      "gatewayMerchantId": "pkey_123"
-                    }
-                  }
-                }
-              ],
-              "transactionInfo": {
-                "totalPrice": "30.00",
-                "totalPriceStatus": "FINAL",
-                "currencyCode": "SGD"
-              },
-              "merchantInfo": {
-                "merchantId": "merchantId"
-              }
-            }
-            """
-        )
-
-        val mapper = ObjectMapper()
-        assertEquals(mapper.readTree(expected.toString()), mapper.readTree(getPaymentDataRequest.toString()))
     }
 }
