+++ sdk/src/test/resources/data/objects/capability_object.json
@@ -46,5 +46,8 @@
       ]
     }
   ],
+  "tokenization_methods": [
+    "googlepay"
+  ],
   "zero_interest_installments": false
 }
